+++ HMI/Export/SqlExportDialog.vb
@@ -10,6 +10,11 @@ Public Class SqlExportDialog
     <ComponentModel.Browsable(False)>
     Public Property SelectedDbmsType As Integer = ExportHelper.DBMS_ORACLE
 
+    ''' <summary>入力されたデータベース名 (USE [DB] 用、空の場合は出力しない)</summary>
+    <ComponentModel.DesignerSerializationVisibility(ComponentModel.DesignerSerializationVisibility.Hidden)>
+    <ComponentModel.Browsable(False)>
+    Public Property DatabaseName As String = ""
+
     Public Sub New()
         InitializeComponent()
         ApplyLocalization()
@@ -23,6 +28,7 @@ Public Class SqlExportDialog
 
         ' ExportOptions の現在値を反映
         chkCreateTable.Checked = ExportOptions.SqlCreateTable
+        chkInferInteger.Checked = ExportOptions.SqlInferInteger
     End Sub
 
     Private Sub btnOK_Click(sender As Object, e As EventArgs) Handles btnOK.Click
@@ -33,8 +39,10 @@ Public Class SqlExportDialog
             Case 3 : SelectedDbmsType = ExportHelper.DBMS_SQLSERVER
         End Select
 
-        ' CREATE TABLE オプションを反映・保存
+        ' オプションを反映・保存
         ExportOptions.SqlCreateTable = chkCreateTable.Checked
+        ExportOptions.SqlInferInteger = chkInferInteger.Checked
+        DatabaseName = txtDatabaseName.Text.Trim()
         ExportOptions.Save()
     End Sub
 
@@ -43,6 +51,8 @@ Public Class SqlExportDialog
         Me.Text = Loc.S("SqlExport_FormTitle")
         lblDbms.Text = Loc.S("SqlExport_DatabaseLabel")
         chkCreateTable.Text = Loc.S("ExportOptions_SqlCreateTable")
+        chkInferInteger.Text = Loc.S("ExportOptions_SqlInferInteger")
+        lblDatabaseName.Text = Loc.S("SqlExport_DatabaseName")
         btnOK.Text = Loc.S("Button_OK")
         btnCancel.Text = Loc.S("Button_Cancel")
     End Sub